trading_bot/requirements.txt
pandas
numpy
scikit-learn
python-binance
matplotlib
pyyaml
